+++ terraform/files/bastion/container_exec.sh
@@ -23,6 +23,6 @@ kubectl get deployment -n \$namespace
 echo -n "Enter deployment name: "
 read dep_name
 
-kubectl exec -it -n \$namespace \$(kubectl get pod -n \$namespace | grep \$dep_name | head -n 1 | cut -d " " -f 1) -- /bin/bash
+kubectl exec -it -n \$namespace \$(kubectl get pod -n \$namespace | grep \$dep_name | grep Running | head -n 1 | cut -d " " -f 1) -- /bin/bash
 echo "Press enter to exit"
 read